create list similar city

+++ lib/model/current_daily_weather.dart
@@ -41,7 +41,7 @@ class CurrentDailyWeather {
     pressure = json['pressure'];
     humidity = json['humidity'];
     dewPoint = json['dew_point'].toDouble();
-    uvi = json['uvi'];
+    uvi = json['uvi'].toDouble();
     clouds = json['clouds'];
     visibility = json['visibility'];
     windSpeed = json['wind_speed'];
